fix send package
